refactor(worker): isolate workspace cleanup lifecycle
