+++ workers/common/agent_runtime.go
@@ -1338,51 +1338,6 @@ func executionWorkspaceCompletionMessage(
 	return fmt.Sprintf("Task %s/%s completed in %s workspace", taskNamespace, taskName, provider)
 }
 
-type executionWorkspaceCleanupOptions struct {
-	skipSubstrateDeleteScrub  bool
-	skipSubstrateReleaseScrub bool
-	statusOptions             []executionWorkspaceStatusOption
-}
-
-func preTerminalExecutionWorkspaceCleanup(
-	workspaceEnv workerenv.ExecutionWorkspaceEnv,
-	substrateHandoffBootstrapped bool,
-	claimedNewWorkspace bool,
-) (workerenv.ExecutionWorkspaceEnv, executionWorkspaceCleanupOptions) {
-	if substrateHandoffBootstrapped ||
-		strings.TrimSpace(workspaceEnv.Provider) != string(corev1alpha1.WorkspaceProviderSubstrate) {
-		return workspaceEnv, executionWorkspaceCleanupOptions{}
-	}
-	options := executionWorkspaceCleanupOptions{
-		skipSubstrateDeleteScrub:  true,
-		skipSubstrateReleaseScrub: true,
-	}
-	if claimedNewWorkspace &&
-		strings.EqualFold(
-			strings.TrimSpace(workspaceEnv.CleanupPolicy),
-			string(corev1alpha1.WorkspaceCleanupPolicyRetain),
-		) {
-		workspaceEnv.CleanupPolicy = string(corev1alpha1.WorkspaceCleanupPolicyDelete)
-	}
-	return workspaceEnv, options
-}
-
-func executionWorkspaceCleanupPolicy(workspaceEnv workerenv.ExecutionWorkspaceEnv) string {
-	policy := corev1alpha1.WorkspaceCleanupPolicy(strings.TrimSpace(strings.ToLower(workspaceEnv.CleanupPolicy)))
-	if strings.TrimSpace(workspaceEnv.Provider) == string(corev1alpha1.WorkspaceProviderSubstrate) &&
-		strings.TrimSpace(workspaceEnv.PoolName) != "" &&
-		policy != "" &&
-		policy != corev1alpha1.WorkspaceCleanupPolicyDelete {
-		return string(corev1alpha1.WorkspaceCleanupPolicyDelete)
-	}
-	return string(policy)
-}
-
-func executionWorkspaceDeferredCleanupSubmitsStatus(workspaceEnv workerenv.ExecutionWorkspaceEnv) bool {
-	return strings.TrimSpace(workspaceEnv.Provider) == string(corev1alpha1.WorkspaceProviderSubstrate) &&
-		strings.TrimSpace(workspaceEnv.PoolName) != ""
-}
-
 func runAgentInSandbox(ctx context.Context, name, workspaceDir string, sandboxEnv workerenv.AgentSandboxEnv) error {
 	executor := getAgentSandboxWorkspaceExecutor()
 	if executor == nil {
@@ -1770,13 +1725,6 @@ func executionWorkspaceResidentKey(workspaceEnv workerenv.ExecutionWorkspaceEnv,
 	return strings.TrimSpace(ref.ClaimName)
 }
 
-func executionWorkspaceCheckpointURI(workspaceEnv workerenv.ExecutionWorkspaceEnv) string {
-	if !workspaceEnv.SnapshotOnRelease {
-		return ""
-	}
-	return strings.TrimSpace(workspaceEnv.SnapshotCheckpointURI)
-}
-
 func stageAgentSandboxExecutable(
 	ctx context.Context,
 	executor workspace.WorkspaceExecutor,
@@ -2023,168 +1971,6 @@ func cleanupAgentSandboxWorkspace(
 		)
 		retainAgentSandboxWorkspace(ctx, executor, ref, sandboxEnv, "unsupported agent sandbox cleanup policy")
 	}
-}
-
-func cleanupExecutionWorkspace(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	workspaceEnv workerenv.ExecutionWorkspaceEnv,
-	reused bool,
-	submitStatus bool,
-) error {
-	return cleanupExecutionWorkspaceWithOptions(
-		ctx,
-		executor,
-		ref,
-		workspaceEnv,
-		reused,
-		submitStatus,
-		executionWorkspaceCleanupOptions{},
-	)
-}
-
-func cleanupExecutionWorkspaceWithOptions(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	workspaceEnv workerenv.ExecutionWorkspaceEnv,
-	reused bool,
-	submitStatus bool,
-	options executionWorkspaceCleanupOptions,
-) error {
-	if ref.IsZero() || executor == nil {
-		return nil
-	}
-
-	cleanupPolicy := executionWorkspaceCleanupPolicy(workspaceEnv)
-	statusEnv := workspaceEnv
-	if cleanupPolicy != "" {
-		statusEnv.CleanupPolicy = cleanupPolicy
-	}
-
-	switch cleanupPolicy {
-	case "retain":
-		if shouldPreScrubExecutionWorkspaceSecrets(workspaceEnv) {
-			if err := scrubExecutionWorkspaceSecrets(ctx, executor, ref, workspaceEnv); err != nil {
-				return fmt.Errorf("%w: %w", errExecutionWorkspaceSecretScrubFailed, err)
-			}
-		}
-		if _, err := executor.Release(ctx, workspace.ReleaseRequest{
-			Ref:                   ref,
-			Retain:                true,
-			Reason:                "execution workspace cleanup policy retain",
-			Timeout:               workspaceEnv.ClaimTimeout,
-			SkipScrub:             options.skipSubstrateReleaseScrub,
-			SnapshotCheckpointURI: executionWorkspaceCheckpointURI(workspaceEnv),
-		}); err != nil {
-			return fmt.Errorf("retain workspace: %w", err)
-		}
-		if submitStatus {
-			submitExecutionWorkspaceStatus(
-				statusEnv,
-				corev1alpha1.ExecutionWorkspacePhaseRetained,
-				corev1alpha1.ExecutionWorkspaceReasonRetained,
-				reused,
-				"workspace retained",
-				options.statusOptions...,
-			)
-		}
-		return nil
-	case "", "delete":
-		if _, err := executor.Delete(ctx, workspace.DeleteRequest{
-			Ref:       ref,
-			Reason:    "execution workspace cleanup policy delete",
-			Timeout:   workspaceEnv.ClaimTimeout,
-			SkipScrub: options.skipSubstrateDeleteScrub,
-		}); err != nil {
-			return fmt.Errorf("delete workspace: %w", err)
-		}
-		if submitStatus {
-			submitExecutionWorkspaceStatus(
-				statusEnv,
-				corev1alpha1.ExecutionWorkspacePhaseDeleted,
-				corev1alpha1.ExecutionWorkspaceReasonDeleted,
-				reused,
-				"workspace deleted",
-				options.statusOptions...,
-			)
-		}
-		return nil
-	default:
-		fmt.Fprintf(
-			os.Stderr,
-			"warning: unsupported workspace cleanup policy %q; retaining workspace to avoid unintended deletion\n",
-			workspaceEnv.CleanupPolicy,
-		)
-		if shouldPreScrubExecutionWorkspaceSecrets(workspaceEnv) {
-			if err := scrubExecutionWorkspaceSecrets(ctx, executor, ref, workspaceEnv); err != nil {
-				return fmt.Errorf("%w: %w", errExecutionWorkspaceSecretScrubFailed, err)
-			}
-		}
-		if _, err := executor.Release(ctx, workspace.ReleaseRequest{
-			Ref:                   ref,
-			Retain:                true,
-			Reason:                "unsupported execution workspace cleanup policy",
-			Timeout:               workspaceEnv.ClaimTimeout,
-			SkipScrub:             options.skipSubstrateReleaseScrub,
-			SnapshotCheckpointURI: executionWorkspaceCheckpointURI(workspaceEnv),
-		}); err != nil {
-			return fmt.Errorf("retain workspace after unsupported cleanup policy: %w", err)
-		}
-		if submitStatus {
-			submitExecutionWorkspaceStatus(
-				statusEnv,
-				corev1alpha1.ExecutionWorkspacePhaseRetained,
-				corev1alpha1.ExecutionWorkspaceReasonRetained,
-				reused,
-				"workspace retained",
-				options.statusOptions...,
-			)
-		}
-		return nil
-	}
-}
-
-func shouldPreScrubExecutionWorkspaceSecrets(workspaceEnv workerenv.ExecutionWorkspaceEnv) bool {
-	return strings.TrimSpace(workspaceEnv.Provider) != string(corev1alpha1.WorkspaceProviderSubstrate)
-}
-
-func scrubExecutionWorkspaceSecrets(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	workspaceEnv workerenv.ExecutionWorkspaceEnv,
-) error {
-	paths := executionWorkspaceScrubPaths()
-	if len(paths) == 0 {
-		return nil
-	}
-	_, err := executor.Exec(ctx, workspace.ExecRequest{
-		Ref:            ref,
-		Command:        append([]string{"rm", "-f"}, paths...),
-		Timeout:        workspaceEnv.ClaimTimeout,
-		MaxOutputBytes: agentSandboxExecMaxOutputBytes,
-	})
-	if err != nil {
-		return fmt.Errorf("scrub execution workspace staged credentials: %w", err)
-	}
-	return nil
-}
-
-func executionWorkspaceScrubPaths() []string {
-	paths := []string{
-		agentSandboxWorkerExecPath,
-		agentSandboxSATokenExecPath,
-		agentSandboxTransactionTokenExecPath,
-		agentSandboxContextSubjectTokenExecPath,
-		agentSandboxGitAskpassExecPath,
-		workspaceHandoffTokenDefaultPath,
-	}
-	if custom := strings.TrimSpace(os.Getenv(workspaceHandoffTokenFileEnv)); custom != "" {
-		paths = appendUniqueString(paths, custom)
-	}
-	return paths
 }
 
 func workspaceHandoffTokenUploadTarget() string {
